requirements.txt
# AI/LLM 프레임워크
langchain-aws==0.2.29
langsmith==0.4.5

# 데이터 검증 & 설정
pydantic-settings==2.10.1
marshmallow==3.26.1
dataclasses-json==0.6.7

# HTTP 클라이언트
httpx==0.28.1
httpx-sse==0.4.1
aiohttp==3.12.14
requests==2.32.4
requests-toolbelt==1.0.0

# AWS
boto3==1.39.14
boto3-stubs==1.39.4

# 비동기
aiohappyeyeballs==2.6.1
anyio==4.9.0
greenlet==3.2.3

# 데이터 처리
numpy==2.3.1
orjson==3.10.18
jsonpatch==1.33
jsonpointer==3.0.0

# 설정 & 환경
python-dotenv==1.1.1
PyYAML==6.0.2

# 테스팅
pytest==8.4.1

# 유틸리티
tenacity==9.1.2
attrs==25.3.0
python-dateutil==2.9.0.post0
Pygments==2.19.2
jmespath==1.0.1
zstandard==0.23.0

# 호환성 & 저수준
typing_extensions==4.14.1
annotated-types==0.7.0
packaging==24.2
certifi==2025.7.9
charset-normalizer==3.4.2
urllib3==2.5.0
idna==3.10
six==1.17.0
sniffio==1.3.1